refactor(desktop): move CommandPaletteTrigger into its own component folder
CommandPaletteHost.tsx held two components; every other mount it renders
alongside (DeleteWorkspaceMount, FolderImportMount, QuickCreateWorkspaceMount,
etc.) already lives in ui/<Name>/<Name>.tsx, making this file the outlier.
Moves CommandPaletteTrigger to ui/CommandPaletteTrigger/ to match, leaving
CommandPaletteHost.tsx as a single component.
