+++ apps/desktop/src/renderer/commandPalette/CommandPaletteHost.tsx
@@ -1,10 +1,8 @@
 import { type ReactNode, useEffect } from "react";
-import { useHotkey } from "renderer/hotkeys";
 import { CommandContextProvider } from "./core/ContextProvider";
-import { useFrameStackStore } from "./core/frames";
-import { useCheckResourcesHotkey } from "./hooks/useCheckResourcesHotkey";
 import { registerAllModules } from "./modules";
 import { CommandPalette } from "./ui/CommandPalette/CommandPalette";
+import { CommandPaletteTrigger } from "./ui/CommandPaletteTrigger/CommandPaletteTrigger";
 import { DeleteWorkspaceMount } from "./ui/DeleteWorkspaceMount/DeleteWorkspaceMount";
 import { FolderImportMount } from "./ui/FolderImportMount/FolderImportMount";
 import { QuickCreateWorkspaceMount } from "./ui/QuickCreateWorkspaceMount/QuickCreateWorkspaceMount";
@@ -29,17 +27,4 @@ export function CommandPaletteHost({ children }: { children?: ReactNode }) {
 			{children}
 		</CommandContextProvider>
 	);
-}
-
-function CommandPaletteTrigger() {
-	const setOpen = useFrameStackStore((s) => s.setOpen);
-	const reset = useFrameStackStore((s) => s.reset);
-	useHotkey("OPEN_COMMAND_PALETTE", () => setOpen(true));
-
-	useCheckResourcesHotkey(() => {
-		setOpen(false);
-		reset();
-	});
-
-	return null;
 }
