+++ apps/desktop/src/renderer/commandPalette/ui/CommandPaletteTrigger/index.ts
@@ -1,0 +1,1 @@
+export { CommandPaletteTrigger } from "./CommandPaletteTrigger";
